Refactor tests and merge cart checkout coupons
